Move legacy frontend bases out of `tests.__init__`
