+++ tests/frontend/test_facilities.py
@@ -2,7 +2,8 @@
 # This file is part of the Pycroft project and licensed under the terms of
 # the Apache License, Version 2.0. See the LICENSE file for details.
 
-from tests import FrontendDataTestBase, FactoryWithConfigDataTestBase, AdminPropertyGroupFactory
+from tests import FactoryWithConfigDataTestBase, AdminPropertyGroupFactory
+from tests.frontend.legacy_base import FrontendDataTestBase
 from tests.factories import RoomFactory
 from tests.factories.user import UserWithMembershipFactory
