+++ tests/frontend/test_permissions.py
@@ -5,8 +5,9 @@
 from flask import url_for, current_app
 from jinja2.runtime import Context
 
-from tests import FrontendDataTestBase, FactoryDataTestBase, \
+from tests import FactoryDataTestBase, \
     FactoryWithConfigDataTestBase, AdminPropertyGroupFactory, MembershipFactory
+from tests.frontend.legacy_base import FrontendDataTestBase
 from tests.factories.property import FinancePropertyGroupFactory
 from tests.factories.user import UserWithMembershipFactory
 from web.template_filters import require
